Update deps, move react to peerDependencies
Without declaring react as a peer dependency, this module will tend to
get moved up to be parallel to tap, causing two copies of ink and
react16 to be installed when tap is used in a project that uses some
other version of react.

Of course, this doesn't much matter for the time being, because tap
bundles treport, ink, and react, but once npm v7 is in wide usage, we
can drop the bundleDependencies kludge, and let the package manager
manage the packages.