fix: Make it work on node v17, update everything
This required a lot more work than I typically like to do in a single
commit.  However, there really was no way to get it all working without
getting it ALL working, so here we are.

React is upgraded to v17, because the only way to silence all
ink-derived security advisories was to upgrade ink, which has a
dependency on a version of react-reconciler than started requiring react
17 in a minor.

This also means that there's entirely new rules about what is allowed to
go in what, the <Color> tag is gone, the <Static> tag has a whole
different interface, and some other changes.  At least Vadim had the
decency to do all this in a SemVer major, which is how it should be
done.

Silencing the ReDOS errors coming from string-length and friends
required upgrading to the newer version of string-length, which is of
course all-in on ESM now, so that stuff has been moved from being a
proper dependency to being a bundled de-ESM-ified version of those deps.
I do not like that, but until we stop supporting pre-ESM versions of
Node (and rewrite tap in ESM, which... yeah, probably not any time
soon?) there's no other way forward that I could see.

Making import-jsx work on node 17 required pulling in my git fork for
the time being, until https://github.com/vadimdemedes/import-jsx/pull/17
lands in the published version.

I think it's probably time to re-think this entire approach, get away
from React and JSX in favor of something a bit less cutting-edge.